moved audio ui Views to their own file
